pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ramp-coremaker"
authors = [
    {name = "Nuclear RAMP team", email = "nuclear.ramp.dev@gmail.com"},
]
description = "Package for representation of geometric and material compositions of nuclear reactor components"
dynamic = ["version"]
requires-python = ">=3.11"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "scipy",
    "numpy",
    "cytoolz",
    "pyparsing",
    "more-itertools",
    "multipledispatch",
    "ramp-core",
    "ramp-isotopes",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-regressions",
    "hypothesis"
]

[tool.setuptools.packages.find]
include = ['coremaker*']
exclude = ['tests*']

[tool.setuptools_scm]
